+++ shared/src/simulation/activities/delay-event.ts
@@ -4,7 +4,7 @@ import { getCreate } from '../../models/utils';
 import { UUID, uuidValidationOptions } from '../../utils';
 import { IsValue } from '../../utils/validators';
 import { ExerciseSimulationEvent, simulationEventTypeOptions } from '../events';
-import { sendSimulationEvent } from '../utils/simulated-region';
+import { sendSimulationEvent } from '../events/utils';
 import type {
     SimulationActivity,
     SimulationActivityState,
@@ -15,15 +15,15 @@ export class DelayEventActivityState implements SimulationActivityState {
     public readonly type = 'delayEventActivity';
 
     @IsUUID(4, uuidValidationOptions)
-    public readonly id!: UUID;
+    public readonly id: UUID;
 
     @Type(...simulationEventTypeOptions)
     @ValidateNested()
-    public readonly event!: ExerciseSimulationEvent;
+    public readonly event: ExerciseSimulationEvent;
 
     @IsInt()
     @Min(0)
-    public readonly endTime!: number;
+    public readonly endTime: number;
 
     /**
      * @deprecated Use {@link create} instead