+++ shared/src/simulation/utils/simulation.ts
@@ -1,12 +1,11 @@
 import type { SimulatedRegion } from '../../models';
 import type { ExerciseState } from '../../state';
-import type { Mutable, UUID } from '../../utils';
-import { cloneDeepMutable } from '../../utils';
-import type { ExerciseSimulationActivityState } from '../activities';
+import type { Mutable } from '../../utils';
 import { simulationActivityDictionary } from '../activities';
+import { terminateActivity } from '../activities/utils';
 import { simulationBehaviorDictionary } from '../behaviors';
-import type { ExerciseSimulationEvent } from '../events';
 import { TickEvent } from '../events/tick';
+import { sendSimulationEvent } from '../events/utils';
 
 export function simulateAllRegions(
     draftState: Mutable<ExerciseState>,
@@ -64,34 +63,4 @@ function handleSimulationEvents(
         });
     });
     simulatedRegion.inEvents = [];
-}
-
-export function sendSimulationEvent(
-    simulatedRegion: Mutable<SimulatedRegion>,
-    event: ExerciseSimulationEvent
-) {
-    simulatedRegion.inEvents.push(cloneDeepMutable(event));
-}
-
-export function addActivity(
-    simulatedRegion: Mutable<SimulatedRegion>,
-    activityState: ExerciseSimulationActivityState
-) {
-    simulatedRegion.activities[activityState.id] =
-        cloneDeepMutable(activityState);
-}
-
-export function terminateActivity(
-    draftState: Mutable<ExerciseState>,
-    simulatedRegion: Mutable<SimulatedRegion>,
-    activityId: UUID
-) {
-    const activityType = simulatedRegion.activities[activityId]?.type;
-    if (activityType) {
-        const activity = simulationActivityDictionary[activityType];
-        if (activity.onTerminate) {
-            activity.onTerminate(draftState, simulatedRegion, activityId);
-        }
-        delete simulatedRegion.activities[activityId];
-    }
 }